refactor(cli): extract `improve` command into cli_improve.py
CI's module-size-limit check failed (cli.py: 1641 lines vs 1600
grandfathered ceiling) after the AC-750..752 additions. Rather than
bump the ceiling, extract the entire `improve` command into a new
`src/autocontext/cli_improve.py` and register it from `cli.py` via the
same `register_*_command(app, console=console)` pattern used by
analytics, hermes, new-scenario, solve, queue, and worker.

cli.py: 1641 -> 1482 lines (back under the 1600 ceiling).
cli_improve.py: 205 lines (under the 800 default limit).

Implementation details:
- `register_improve_command` mirrors `register_analytics_command`:
  takes `(app, *, console, dependency_module="autocontext.cli")` so the
  command can reach `load_settings`, `_write_json_stdout`, and
  `_exit_provider_error` on the host CLI module via `_cli_attr` without
  introducing a circular import.
- All existing test patches keep working because the patched symbols
  (load_settings, ImprovementLoop, get_provider) are still resolved on
  their original module paths.

All 105 CLI tests + module-size-limit test pass. Ruff and mypy clean.
