+++ autocontext/src/autocontext/cli.py
@@ -8,7 +8,6 @@ import sys
 import threading
 import time
 import uuid
-from collections.abc import Callable
 from dataclasses import dataclass
 from pathlib import Path
 from typing import TYPE_CHECKING, Any, NoReturn
@@ -21,6 +20,7 @@ from rich.table import Table
 from autocontext.agents.orchestrator import AgentOrchestrator
 from autocontext.cli_analytics import register_analytics_command
 from autocontext.cli_hermes import register_hermes_command
+from autocontext.cli_improve import register_improve_command
 from autocontext.cli_investigate import run_investigate_command
 from autocontext.cli_new_scenario import register_new_scenario_command
 from autocontext.cli_queue import register_queue_command
@@ -1470,167 +1470,9 @@ def judge(
         console.print(f"[bold]Reasoning:[/bold] {result.reasoning}")
 
 
-@app.command()
-def improve(
-    task_prompt: str = typer.Option(..., "--task-prompt", "-p", help="The task prompt"),
-    rubric: str = typer.Option(..., "--rubric", "-r", help="Evaluation rubric"),
-    initial_output: str = typer.Option("", "--output", "-o", help="Starting output to improve"),
-    max_rounds: int = typer.Option(5, "--rounds", "-n", help="Maximum improvement rounds"),
-    threshold: float = typer.Option(0.9, "--threshold", "-t", help="Quality threshold to stop"),
-    provider_override: str = typer.Option("", "--provider", help="Provider override"),
-    timeout: float | None = typer.Option(
-        None,
-        "--timeout",
-        min=1.0,
-        help=(
-            "Override per-call provider timeout in seconds. For claude-cli this "
-            "writes claude_timeout (env: AUTOCONTEXT_CLAUDE_TIMEOUT, default 600s); "
-            "for codex it writes codex_timeout; for pi/pi-rpc it writes pi_timeout. "
-            "For the overall claude-cli wall-clock budget, see --claude-max-total-seconds."
-        ),
-    ),
-    claude_max_total_seconds: float | None = typer.Option(
-        None,
-        "--claude-max-total-seconds",
-        min=0.0,
-        help=(
-            "Override the wall-clock ceiling on total claude-cli runtime across all "
-            "invocations during this run (env: AUTOCONTEXT_CLAUDE_MAX_TOTAL_SECONDS, "
-            "default 0=off). Only applied when the resolved judge provider is claude-cli."
-        ),
-    ),
-    json_output: bool = typer.Option(False, "--json", help="Output structured JSON"),
-    ndjson_output: bool = typer.Option(
-        False,
-        "--ndjson",
-        help=(
-            "Stream per-round events as newline-delimited JSON to stdout (AC-752). "
-            "Useful for long-running loops where --json would buffer all output until "
-            "completion. Emits one JSON line per event: round_start, judge_done, "
-            "verifier_done, round_summary, and a final summary line."
-        ),
-    ),
-    verify_cmd: str = typer.Option(
-        "",
-        "--verify-cmd",
-        help=(
-            "External command to verify each round's output (AC-733). "
-            "Non-zero exit forces the round score to 0 and feeds the "
-            "command's stderr/stdout into the next revision prompt. "
-            "Use the literal `{file}` placeholder to receive the output as a "
-            "temp-file path; otherwise the output is piped to stdin. "
-            "Examples: 'lake env lean {file}', 'mypy {file}', 'cargo check'."
-        ),
-    ),
-    verify_suffix: str = typer.Option(
-        ".txt",
-        "--verify-suffix",
-        help="Suffix for the temp file passed to --verify-cmd (e.g. '.lean', '.py').",
-    ),
-    verify_timeout: float = typer.Option(
-        300.0,
-        "--verify-timeout",
-        min=1.0,
-        help="Timeout in seconds for each --verify-cmd invocation.",
-    ),
-) -> None:
-    """Run multi-round improvement loop on agent output.
-
-    Creates a simple agent task from the prompt and rubric, then runs
-    the improvement loop with judge-guided iteration.
-    """
-
-    from autocontext.execution.improvement_events import ImprovementLoopEvent
-    from autocontext.execution.improvement_loop import ImprovementLoop
-    from autocontext.execution.output_verifier import make_verifier
-    from autocontext.execution.task_runner import SimpleAgentTask
-    from autocontext.providers.registry import get_provider as get_judge_provider
-
-    # AC-752 (P3 follow-up): --json (single final blob) and --ndjson (streaming
-    # events) are mutually exclusive output modes. Passing both produces a
-    # mixed, un-parseable stream. Reject up front with a clear error.
-    if json_output and ndjson_output:
-        typer.echo(
-            "Error: --json and --ndjson are mutually exclusive output modes; pick one.",
-            err=True,
-        )
-        raise typer.Exit(code=2)
-
-    settings = apply_judge_runtime_overrides(
-        load_settings(),
-        provider_name=provider_override,
-        timeout=timeout,
-        claude_max_total_seconds=claude_max_total_seconds,
-    )
-
-    try:
-        provider = get_judge_provider(settings)
-        task = SimpleAgentTask(
-            task_prompt=task_prompt,
-            rubric=rubric,
-            provider=provider,
-            model=settings.judge_model,
-        )
-        state = task.initial_state()
-        verifier = make_verifier(
-            verify_cmd or None,
-            file_suffix=verify_suffix,
-            timeout_s=verify_timeout,
-        )
-        # AC-752: when --ndjson is set, stream per-round events as JSON lines
-        # so long-running loops have progress visibility before --json's final
-        # blob lands. The event sink writes one compact JSON line per event.
-        on_event: Callable[[ImprovementLoopEvent], None] | None = None
-        if ndjson_output:
-
-            def _emit_ndjson(event: ImprovementLoopEvent) -> None:
-                payload = {k: v for k, v in dataclasses.asdict(event).items() if v is not None}
-                typer.echo(json.dumps(payload))
-
-            on_event = _emit_ndjson
-        loop = ImprovementLoop(
-            task=task,
-            max_rounds=max_rounds,
-            quality_threshold=threshold,
-            output_verifier=verifier,
-            on_event=on_event,
-        )
-        starting_output = initial_output or task.generate_output(state)
-        result = loop.run(initial_output=starting_output, state=state)
-    except ProviderError as exc:
-        _exit_provider_error(
-            exc,
-            provider_name=settings.judge_provider,
-            settings=settings,
-            json_output=json_output,
-            ndjson_output=ndjson_output,
-        )
-
-    if ndjson_output:
-        # AC-752: under --ndjson, stdout is pure newline-delimited JSON
-        # (already streamed via on_event). Suppress the Rich human-readable
-        # summary so consumers can reliably parse each stdout line as JSON.
-        # --json + --ndjson is rejected up front, so we don't need to handle
-        # the both-set case here.
-        pass
-    elif json_output:
-        _write_json_stdout(
-            {
-                "best_score": result.best_score,
-                "best_round": result.best_round,
-                "total_rounds": result.total_rounds,
-                "met_threshold": result.met_threshold,
-                "best_output": result.best_output,
-            }
-        )
-    else:
-        console.print(f"[bold]Best score:[/bold] {result.best_score:.4f} (round {result.best_round})")
-        console.print(f"[bold]Rounds:[/bold] {result.total_rounds}")
-        console.print(f"[bold]Met threshold:[/bold] {result.met_threshold}")
-
-
 register_analytics_command(app, console=console)
 register_hermes_command(app, console=console)
+register_improve_command(app, console=console)
 register_new_scenario_command(app, console=console)
 register_solve_command(app, console=console)
 register_queue_command(app, console=console)
